portfolio_backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
Pillow
python-dotenv
